+++ typescript/infra/config/environments/mainnet3/warp/configGetters/getoUSDTTokenWarpConfig.ts
@@ -32,15 +32,11 @@ export const deploymentChains = [
   'superseed',
   'lisk',
   'worldchain',
-  'sonic',
-  'bitlayer',
-  'ronin',
   'mantle',
   'metis',
   'linea',
   'metal',
   'bob',
-  'hashkey',
   'swell',
   'zerogravity',
 ] as const;
@@ -71,15 +67,11 @@ const productionBufferCapByChain: TypedoUSDTTokenChainMap<string> = {
   superseed: lowerBufferCap,
   lisk: lowerBufferCap,
   worldchain: '0',
-  sonic: middleBufferCap,
-  bitlayer: lowerBufferCap,
-  ronin: lowerBufferCap,
   mantle: middleBufferCap,
   metis: lowerBufferCap,
   linea: lowerBufferCap,
   metal: lowerBufferCap,
   bob: lowerBufferCap,
-  hashkey: lowerBufferCap,
   swell: middleBufferCap,
   zerogravity: middleBufferCap,
 };
@@ -99,15 +91,11 @@ const productionRateLimitByChain: TypedoUSDTTokenChainMap<string> = {
   superseed: lowerRateLimitPerSecond,
   lisk: lowerRateLimitPerSecond,
   worldchain: '0',
-  sonic: middleRateLimitPerSecond,
-  bitlayer: lowerRateLimitPerSecond,
-  ronin: lowerRateLimitPerSecond,
   mantle: middleRateLimitPerSecond,
   metis: lowerRateLimitPerSecond,
   linea: lowerRateLimitPerSecond,
   metal: lowerRateLimitPerSecond,
   bob: lowerRateLimitPerSecond,
-  hashkey: lowerRateLimitPerSecond,
   swell: middleRateLimitPerSecond,
   zerogravity: middleRateLimitPerSecond,
 };
@@ -176,18 +164,6 @@ const productionOwnerOverridesByChain: TypedoUSDTTokenChainMap<
     collateralToken: productionOwnerByChain.worldchain,
     collateralProxyAdmin: productionOwnerByChain.worldchain,
   },
-  sonic: {
-    collateralToken: productionOwnerByChain.sonic,
-    collateralProxyAdmin: productionOwnerByChain.sonic,
-  },
-  bitlayer: {
-    collateralToken: productionOwnerByChain.bitlayer,
-    collateralProxyAdmin: productionOwnerByChain.bitlayer,
-  },
-  ronin: {
-    collateralToken: productionOwnerByChain.ronin,
-    collateralProxyAdmin: productionOwnerByChain.ronin,
-  },
   mantle: {
     collateralToken: productionOwnerByChain.mantle,
     collateralProxyAdmin: productionOwnerByChain.mantle,
@@ -207,10 +183,6 @@ const productionOwnerOverridesByChain: TypedoUSDTTokenChainMap<
   bob: {
     collateralToken: productionOwnerByChain.bob,
     collateralProxyAdmin: productionOwnerByChain.bob,
-  },
-  hashkey: {
-    collateralToken: productionOwnerByChain.hashkey,
-    collateralProxyAdmin: productionOwnerByChain.hashkey,
   },
   swell: {
     collateralToken: productionOwnerByChain.swell,
@@ -240,10 +212,8 @@ const productionCCIPTokenPoolAddresses: ChainMap<Address> = {
   ethereum: '0xa3532633401AbFfbd15e6be825a45FB7F141469B',
   celo: '0x47Db76c9c97F4bcFd54D8872FDb848Cab696092d',
   base: '0xa760D20a91C076A57b270D3F7a3150421ab40591',
-  sonic: '0x6a21a19aD44542d83F7f7FF45Aa31A62a36200de',
   optimism: '0x6a21a19aD44542d83F7f7FF45Aa31A62a36200de',
   bob: '0xAFEd606Bd2CAb6983fC6F10167c98aaC2173D77f',
-  hashkey: '0x55aeb80Aa6Ab34aA83E1F387903F8Bb2Aa9e2F2d',
   zerogravity: '0xd7502CaBdb70c79382deF58FB6df3CdA69cb2A1b',
 };
 
@@ -290,12 +260,6 @@ const productionExtraBridges: ChainMap<XERC20TokenExtraBridgesLimits[]> = {
       limits: productionCCIPTokenPoolLimits,
     },
   ],
-  sonic: [
-    {
-      lockbox: productionCCIPTokenPoolAddresses.sonic,
-      limits: productionCCIPTokenPoolLimits,
-    },
-  ],
   optimism: [
     {
       // usdc
@@ -315,12 +279,6 @@ const productionExtraBridges: ChainMap<XERC20TokenExtraBridgesLimits[]> = {
   bob: [
     {
       lockbox: productionCCIPTokenPoolAddresses.bob,
-      limits: productionCCIPTokenPoolLimits,
-    },
-  ],
-  hashkey: [
-    {
-      lockbox: productionCCIPTokenPoolAddresses.hashkey,
       limits: productionCCIPTokenPoolLimits,
     },
   ],
@@ -345,15 +303,11 @@ const productionXERC20AddressesByChain: TypedoUSDTTokenChainMap<Address> = {
   superseed: productionXERC20TokenAddress,
   lisk: productionXERC20TokenAddress,
   worldchain: productionXERC20TokenAddress,
-  sonic: productionXERC20TokenAddress,
-  bitlayer: productionXERC20TokenAddress,
-  ronin: productionXERC20TokenAddress,
   mantle: productionXERC20TokenAddress,
   metis: productionXERC20TokenAddress,
   linea: productionXERC20TokenAddress,
   metal: productionXERC20TokenAddress,
   bob: productionXERC20TokenAddress,
-  hashkey: productionXERC20TokenAddress,
   swell: productionXERC20TokenAddress,
   zerogravity: productionXERC20TokenAddress,
 };
@@ -373,15 +327,11 @@ const stagingBufferCapByChain: TypedoUSDTTokenChainMap<string> = {
   superseed: stagingDefaultBufferCap,
   lisk: stagingDefaultBufferCap,
   worldchain: stagingDefaultBufferCap,
-  sonic: stagingDefaultBufferCap,
-  bitlayer: stagingDefaultBufferCap,
-  ronin: stagingDefaultBufferCap,
   mantle: stagingDefaultBufferCap,
   metis: stagingDefaultBufferCap,
   linea: stagingDefaultBufferCap,
   metal: stagingDefaultBufferCap,
   bob: stagingDefaultBufferCap,
-  hashkey: stagingDefaultBufferCap,
   swell: stagingDefaultBufferCap,
   zerogravity: stagingDefaultBufferCap,
 };
@@ -399,15 +349,11 @@ const stagingRateLimitByChain: TypedoUSDTTokenChainMap<string> = {
   superseed: stagingDefaultRateLimitPerSecond,
   lisk: stagingDefaultRateLimitPerSecond,
   worldchain: stagingDefaultRateLimitPerSecond,
-  sonic: stagingDefaultRateLimitPerSecond,
-  bitlayer: stagingDefaultRateLimitPerSecond,
-  ronin: stagingDefaultRateLimitPerSecond,
   mantle: stagingDefaultRateLimitPerSecond,
   metis: stagingDefaultRateLimitPerSecond,
   linea: stagingDefaultRateLimitPerSecond,
   metal: stagingDefaultRateLimitPerSecond,
   bob: stagingDefaultRateLimitPerSecond,
-  hashkey: stagingDefaultRateLimitPerSecond,
   swell: stagingDefaultRateLimitPerSecond,
   zerogravity: stagingDefaultRateLimitPerSecond,
 };
@@ -437,15 +383,11 @@ const stagingXERC20AddressesByChain: TypedoUSDTTokenChainMap<Address> = {
   superseed: stagingXERC20TokenAddress,
   lisk: stagingXERC20TokenAddress,
   worldchain: stagingXERC20TokenAddress,
-  sonic: stagingXERC20TokenAddress,
-  bitlayer: stagingXERC20TokenAddress,
-  ronin: stagingXERC20TokenAddress,
   mantle: stagingXERC20TokenAddress,
   metis: stagingXERC20TokenAddress,
   linea: stagingXERC20TokenAddress,
   metal: stagingXERC20TokenAddress,
   bob: stagingXERC20TokenAddress,
-  hashkey: stagingXERC20TokenAddress,
   swell: stagingXERC20TokenAddress,
   zerogravity: stagingXERC20TokenAddress,
 };