+++ typescript/fee-quoting/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @hyperlane-xyz/fee-quoting
 
+## 27.3.8
+
+### Patch Changes
+
+- Updated dependencies [4a21153]
+- Updated dependencies [c328efa]
+- Updated dependencies [4a21153]
+- Updated dependencies [4a21153]
+- Updated dependencies [322a418]
+- Updated dependencies [d5d7d1a]
+  - @hyperlane-xyz/sdk@41.3.0
+  - @hyperlane-xyz/sealevel-sdk@41.3.0
+  - @hyperlane-xyz/utils@41.3.0
+  - @hyperlane-xyz/provider-sdk@8.1.1
+
 ## 27.3.7
 
 ### Patch Changes